refactor: use correct namespace
