+++ Assets/EditorTests/Gameplay/WorldEntityTest.cs
@@ -1,5 +1,6 @@
 using System;
 using Balloondle.Gameplay;
+using Balloondle.Gameplay.World;
 using NUnit.Framework;
 using UnityEngine;
