+++ Assets/Source/Gameplay/CharacterInitializer.cs
@@ -1,3 +1,4 @@
+using Balloondle.Gameplay.World;
 using UnityEngine;
 
 namespace Balloondle.Gameplay
