+++ Assets/Source/Gameplay/Rope2DDestructorForWorldEntity.cs
@@ -1,4 +1,5 @@
 ﻿using Balloondle.Gameplay.Physics2D;
+using Balloondle.Gameplay.World;
 
 namespace Balloondle.Gameplay
 {
