+++ Assets/Source/Gameplay/Rope2DSpawner.cs
@@ -1,5 +1,6 @@
 using System;
 using Balloondle.Gameplay.Physics2D;
+using Balloondle.Gameplay.World;
 using UnityEngine;
 
 namespace Balloondle.Gameplay
