+++ Assets/Source/Gameplay/World/Player.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using UnityEngine;
 
-namespace Balloondle.Gameplay
+namespace Balloondle.Gameplay.World
 {
     public class Player : MonoBehaviour
     {
