+++ Assets/Source/Gameplay/World/Rope2DEntityAttacher.cs
@@ -4,7 +4,7 @@ using Balloondle.Gameplay.Physics2D;
 using UnityEngine;
 using UnityEngine.Serialization;
 
-namespace Balloondle.Gameplay
+namespace Balloondle.Gameplay.World
 {
     [CreateAssetMenu(fileName = "Rope2DEntityAttacher", menuName = "WorldEntities/Rope2DEntityAttacher", order = 1)]
     public class Rope2DEntityAttacher : WorldEntityAttacher
